requirements.txt
sqlalchemy
psycopg2-binary
requests
python-dotenv
groq